# backend/openapi.yaml
openapi: 3.1.0
info:
  title: School Management System API
  version: 1.0.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        secondName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, instructor]
        comments:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        taughtCourses:
          type: array
          items:
            $ref: "#/components/schemas/CourseSummary"
        enrollments:
          type: array
          items:
            $ref: "#/components/schemas/EnrollmentSummary"
...
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        secondName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, instructor]
        comments:
          type: string
          nullable: true
    CourseSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        language:
          type: string
          enum: [English, Finnish, Swedish, Russian, German, French]
        level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        material:
          type: string
        instructorId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/PublicUser"
        course:
          $ref: "#/components/schemas/CourseSummary"
    EnrollmentSummary:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
paths:
...
  /users:
    get:
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        
        "401":
          description: Unauthorized

        "500":
          description: Server error
    
    post:
      summary: Create user (instructor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, secondName, email, password]
              properties:
                firstName:
                  type: string
                secondName:
                  type: string
                email:
                  type: string
                password:
                  type: string
                comments:
                  type: string

      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          decription: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/me:
    get:
      summary: Get logged-in user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged-in user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        
        "401":
          description: Unauthorized

        "500":
          description: Server error
...
  /users/:id:
    patch:
      summary: Update user information (instructor only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                secondName:
                  type: string
                email:
                  type: string
                comments:
                  type: string
      
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        
        "400":
          decription: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          decription: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
       

    delete:
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string

        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          decription: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /courses/{id}/enroll:
    post:
      summary: Create enrollment (instructor only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: integer
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Enrollment"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "409":
          description: Student already enrolled in this course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /enrollments:
    get:
      summary: Get all enrollments (instructor only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Enrollment"
        
        "401":
          description: Unauthorized

        "403":
          description: Forbidden

        "500":
          description: Server error
  /enrollments/{id}:
    delete:
      summary: Delete enrollment (instructor only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Enrollment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string

        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"